hiconn/experiments/requirements.txt
boto==2.38.0
paramiko==1.15.2
locustio
fabric
prettytable
gevent
